processdata/wordcloud_ciyun/requirements.txt
pymysql==1.1.0
numpy==1.24.3
pandas==2.0.3
wordcloud==1.9.2
matplotlib==3.7.2
Pillow==10.0.0
jieba==0.42.1